Refactor (#6528)

+++ ct/pairdrop.sh
@@ -27,13 +27,28 @@ function update_script() {
         msg_error "No ${APP} Installation Found!"
         exit
     fi
-    msg_info "Updating $APP"
-    systemctl stop pairdrop
-    cd /opt/pairdrop
-    git pull
-    npm install
-    systemctl start pairdrop
-    msg_ok "Updated $APP"
+
+    RELEASE=$(curl -fsSL https://api.github.com/repos/schlagmichdoch/PairDrop/releases/latest | grep "tag_name" | awk '{print substr($2, 3, length($2)-4) }')
+    if [[ ! -f ~/.pairdrop ]] || [[ "${RELEASE}" != "$(cat ~/.pairdrop)" ]]; then
+      msg_info "Stopping ${APP}"
+      systemctl stop pairdrop
+      msg_ok "Stopped ${APP}"
+      
+      fetch_and_deploy_gh_release "pairdrop" "schlagmichdoch/PairDrop" "tarball"
+
+      msg_info "Configuring PairDrop"
+      cd /opt/pairdrop
+      $STD npm install
+      msg_ok "Configured PairDrop"
+
+      msg_info "Starting ${APP}"
+      systemctl start pairdrop
+      msg_ok "Started ${APP}"
+
+      msg_ok "Updated Successfully"
+    else
+      msg_ok "No update required. ${APP} is already at ${RELEASE}."
+    fi
     exit
 }
